+++ src/packet/payload.rs
@@ -6,7 +6,7 @@ use crate::types::branch;
 
 use super::decoder::{Decode, Decoder};
 use super::encoder::{Encode, Encoder};
-use super::{Error, sync, unit, util};
+use super::{Error, ext, sync, unit, util};
 
 /// An E-Trace payload
 #[derive(Copy, Clone, Debug, PartialEq)]
@@ -46,53 +46,10 @@ impl<I, D> TryFrom<Payload<I, D>> for InstructionTrace<I, D> {
     }
 }
 
-/// Determines the layout of [`BranchCount`].
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum BranchFmt {
-    /// No address
-    ///
-    /// The packet does not contain an address, and the branch following the
-    /// last correct prediction failed.
-    NoAddr = 0,
-    /// Address, success
-    ///
-    /// The packet contains an address. If this points to a branch instruction,
-    /// then the branch was predicted correctly.
-    Addr = 2,
-    /// Address, failure
-    ///
-    /// The packet contains an address that points to a branch instruction. The
-    /// prediction for that branch failed.
-    AddrFail = 3,
-}
-
-impl<U> Decode<'_, U> for BranchFmt {
-    fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
-        match decoder.read_bits::<u8>(2)? {
-            0b00 => Ok(BranchFmt::NoAddr),
-            0b01 => Err(Error::BadBranchFmt),
-            0b10 => Ok(BranchFmt::Addr),
-            0b11 => Ok(BranchFmt::AddrFail),
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl<U> Encode<'_, U> for BranchFmt {
-    fn encode(&self, encoder: &mut Encoder<U>) -> Result<(), Error> {
-        let value: u8 = match self {
-            Self::NoAddr => 0b00,
-            Self::Addr => 0b10,
-            Self::AddrFail => 0b11,
-        };
-        encoder.write_bits(value, 2)
-    }
-}
-
 /// An instruction trace payload
 #[derive(Copy, Clone, Debug, Eq, PartialEq)]
 pub enum InstructionTrace<I = unit::ReferenceIOptions, D = unit::ReferenceDOptions> {
-    Extension(Extension),
+    Extension(ext::Extension),
     Branch(Branch),
     Address(AddressInfo),
     Synchronization(sync::Synchronization<I, D>),
@@ -101,7 +58,7 @@ pub enum InstructionTrace<I = unit::ReferenceIOptions, D = unit::ReferenceDOptio
 impl<U: unit::Unit> Decode<'_, U> for InstructionTrace<U::IOptions, U::DOptions> {
     fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
         match decoder.read_bits::<u8>(2)? {
-            0b00 => Extension::decode(decoder).map(Into::into),
+            0b00 => ext::Extension::decode(decoder).map(Into::into),
             0b01 => Branch::decode(decoder).map(Into::into),
             0b10 => AddressInfo::decode(decoder).map(Into::into),
             0b11 => sync::Synchronization::decode(decoder).map(Into::into),
@@ -147,7 +104,7 @@ impl<I, D> InstructionTrace<I, D> {
         match self {
             Self::Address(addr) => Some(addr),
             Self::Branch(branch) => branch.address.as_ref(),
-            Self::Extension(Extension::BranchCount(branch_count)) => branch_count.address.as_ref(),
+            Self::Extension(e) => e.get_address_info(),
             _ => None,
         }
     }
@@ -169,28 +126,27 @@ impl<I, D> InstructionTrace<I, D> {
         match self {
             Self::Address(a) => a.irdepth,
             Self::Branch(b) => b.address.and_then(|a| a.irdepth),
-            Self::Extension(Extension::BranchCount(b)) => b.address.and_then(|a| a.irdepth),
-            Self::Extension(Extension::JumpTargetIndex(j)) => j.irdepth,
+            Self::Extension(e) => e.implicit_return_depth(),
             _ => None,
         }
     }
 }
 
-impl<I, D> From<Extension> for InstructionTrace<I, D> {
-    fn from(ex: Extension) -> Self {
+impl<I, D> From<ext::Extension> for InstructionTrace<I, D> {
+    fn from(ex: ext::Extension) -> Self {
         Self::Extension(ex)
     }
 }
 
-impl<I, D> From<BranchCount> for InstructionTrace<I, D> {
-    fn from(count: BranchCount) -> Self {
-        Self::Extension(Extension::BranchCount(count))
+impl<I, D> From<ext::BranchCount> for InstructionTrace<I, D> {
+    fn from(count: ext::BranchCount) -> Self {
+        Self::Extension(ext::Extension::BranchCount(count))
     }
 }
 
-impl<I, D> From<JumpTargetIndex> for InstructionTrace<I, D> {
-    fn from(idx: JumpTargetIndex) -> Self {
-        Self::Extension(Extension::JumpTargetIndex(idx))
+impl<I, D> From<ext::JumpTargetIndex> for InstructionTrace<I, D> {
+    fn from(idx: ext::JumpTargetIndex) -> Self {
+        Self::Extension(ext::Extension::JumpTargetIndex(idx))
     }
 }
 
@@ -233,110 +189,6 @@ impl<I, D> From<sync::Context> for InstructionTrace<I, D> {
 impl<I, D> From<sync::Support<I, D>> for InstructionTrace<I, D> {
     fn from(support: sync::Support<I, D>) -> Self {
         Self::Synchronization(support.into())
-    }
-}
-
-/// Extension payload
-///
-/// Represents a format 0 packet.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub enum Extension {
-    BranchCount(BranchCount),
-    JumpTargetIndex(JumpTargetIndex),
-}
-
-impl<U> Decode<'_, U> for Extension {
-    fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
-        match decoder.read_bits(decoder.widths().format0_subformat)? {
-            0 => BranchCount::decode(decoder).map(Self::BranchCount),
-            1 => JumpTargetIndex::decode(decoder).map(Self::JumpTargetIndex),
-            s => Err(Error::UnknownFmt(0, Some(s))),
-        }
-    }
-}
-
-impl<U> Encode<'_, U> for Extension {
-    fn encode(&self, encoder: &mut Encoder<U>) -> Result<(), Error> {
-        match self {
-            Self::BranchCount(branch) => encoder.encode(branch),
-            Self::JumpTargetIndex(jti) => encoder.encode(jti),
-        }
-    }
-}
-
-/// Branch count payload
-///
-/// Represents a format 0, subformat 0 packet. It informs about the number of
-/// correctly predicted branches.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct BranchCount {
-    /// Count of the number of correctly predicted branches, minus 31.
-    pub branch_count: u32,
-    pub branch_fmt: BranchFmt,
-    pub address: Option<AddressInfo>,
-}
-
-impl<U> Decode<'_, U> for BranchCount {
-    fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
-        let branch_count = decoder.read_bits::<u32>(32)? - 31;
-        let branch_fmt = BranchFmt::decode(decoder)?;
-        let address = if branch_fmt == BranchFmt::NoAddr {
-            None
-        } else {
-            Some(AddressInfo::decode(decoder)?)
-        };
-        Ok(BranchCount {
-            branch_count,
-            address,
-            branch_fmt,
-        })
-    }
-}
-
-impl<U> Encode<'_, U> for BranchCount {
-    fn encode(&self, encoder: &mut Encoder<U>) -> Result<(), Error> {
-        encoder.write_bits(self.branch_count + 31, 32)?;
-        encoder.encode(&self.branch_fmt)?;
-        if let Some(address) = self.address.as_ref() {
-            encoder.encode(address)?;
-        }
-        Ok(())
-    }
-}
-
-/// Jump target index payload
-///
-/// Represents a format 0, subformat 1 packet.
-#[derive(Copy, Clone, Debug, Eq, PartialEq)]
-pub struct JumpTargetIndex {
-    /// Index of entry containing the jump's target address
-    pub index: usize,
-    pub branch_map: branch::Map,
-
-    /// Implicit return depth
-    pub irdepth: Option<usize>,
-}
-
-impl<U> Decode<'_, U> for JumpTargetIndex {
-    fn decode(decoder: &mut Decoder<U>) -> Result<Self, Error> {
-        let index = decoder.read_bits(decoder.widths().cache_index)?;
-        let branch_map = util::BranchCount::decode(decoder)?.read_branch_map(decoder)?;
-        let irdepth = util::read_implicit_return(decoder)?;
-        Ok(JumpTargetIndex {
-            index,
-            branch_map,
-            irdepth,
-        })
-    }
-}
-
-impl<U> Encode<'_, U> for JumpTargetIndex {
-    fn encode(&self, encoder: &mut Encoder<U>) -> Result<(), Error> {
-        encoder.write_bits(self.index, encoder.widths().cache_index)?;
-        let count = util::BranchCount(self.branch_map.count());
-        encoder.encode(&count)?;
-        encoder.write_bits(self.branch_map.raw_map(), count.field_length())?;
-        util::write_implicit_return(encoder, self.irdepth)
     }
 }